Little buggers slipping through
Tweaked the image removal algo a bit because of how the layout algo
works. Based on the layout algo, small images tend to float up, so they
don't pass the inspection line.

Well, tho it doesn't make too much sense, the new algo will remove from
the back, but it's selection process will be based on postion. So even
if the little buggers are above the line, if their ticket is called,
they're going out.

Unfiar? Perhaps. But we're trying to keep the set fresh, so... old
images gotta go.